+++ src/predicates/position/in_front_of_predicate.h
@@ -5,7 +5,7 @@
 #ifndef ENV_MODEL_IN_FRONT_OF_PREDICATE_H
 #define ENV_MODEL_IN_FRONT_OF_PREDICATE_H
 
-#include "predicate.h"
+#include "predicates/predicate.h"
 
 class InFrontOfPredicate : public Predicate {
 public:
@@ -19,7 +19,7 @@ public:
     bool
     booleanEvaluation(double lonPositionK, double lonPositionP);
 
-    bool booleanEvaluation(int timeStep,
+    bool booleanEvaluation(size_t timeStep,
                            const std::shared_ptr<World> &world,
                            const std::shared_ptr<Obstacle> &obstacleK,
                            const std::shared_ptr<Obstacle> &obstacleP) override;
@@ -33,7 +33,7 @@ public:
      * @param obstacleP The pth obstacle. This is an optional parameter.
      * @return Constraints defined by the predicate.
      */
-    Constraint constraintEvaluation(int timeStep,
+    Constraint constraintEvaluation(size_t timeStep,
                                     const std::shared_ptr<World>& world,
                                     const std::shared_ptr<Obstacle> &obstacleK,
                                     const std::shared_ptr<Obstacle> &obstacleP) override;
@@ -56,7 +56,7 @@ public:
      * @param obstacleP The pth obstacle. This is an optional parameter.
      * @return Real value indicating robustness of the predicate.
      */
-    double robustEvaluation(int timeStep,
+    double robustEvaluation(size_t timeStep,
                             const std::shared_ptr<World>& world,
                             const std::shared_ptr<Obstacle> &obstacleK,
                             const std::shared_ptr<Obstacle> &obstacleP) override;
