+++ tests/envModelTests/obstacle/test_obstacle.h
@@ -20,11 +20,11 @@ protected:
     double reactionTimeObstacleOne;
     double lengthObstacleOne;
     double widthObstacleOne;
-    std::map<int, std::shared_ptr<Lane>> occupiedLaneObstacleOne{};
-    std::map<int, std::shared_ptr<State>> trajectoryPredictionObstacleOne{};
-    std::map<int, std::shared_ptr<State>> historyObstacleOne{};
+    std::map<size_t, std::shared_ptr<Lane>> occupiedLaneObstacleOne{};
+    std::map<size_t, std::shared_ptr<State>> trajectoryPredictionObstacleOne{};
+    std::map<size_t, std::shared_ptr<State>> historyObstacleOne{};
     Rectangle geoShapeObstacleOne;
-    std::map<int, std::vector<std::shared_ptr<Lanelet>>> occupiedLaneletsObstacleOne{};
+    std::map<size_t, std::vector<std::shared_ptr<Lanelet>>> occupiedLaneletsObstacleOne{};
     std::shared_ptr<Obstacle> obstacleOne;
 
     int idObstacleTwo;
@@ -37,7 +37,7 @@ protected:
     double reactionTimeObstacleTwo;
     double lengthObstacleTwo;
     double widthObstacleTwo;
-    std::map<int, std::shared_ptr<State>> trajectoryPredictionObstacleTwo{};
+    std::map<size_t, std::shared_ptr<State>> trajectoryPredictionObstacleTwo{};
     std::shared_ptr<Obstacle> obstacleTwo;
 
     std::vector<std::shared_ptr<Obstacle>> obstacleList{};
