+++ tests/envModelTests/predicates/test_safe_distance_predicate.cpp
@@ -11,10 +11,10 @@ void SafeDistancePredicateTest::SetUp() {
     std::shared_ptr<State> stateEgoOne = std::make_shared<State>(1, 20, 0, 20, 0, 0, 0, 20, 0);
     std::shared_ptr<State> stateOtherOne = std::make_shared<State>(1, 30, 0, 0, 0, 0, 0, 30, 0);
 
-    std::map<int, std::shared_ptr<State>> trajectoryPredictionEgoVehicle{
+    std::map<size_t, std::shared_ptr<State>> trajectoryPredictionEgoVehicle{
             std::pair<int, std::shared_ptr<State>>(0, stateEgoZero),
             std::pair<int, std::shared_ptr<State>>(1, stateEgoOne)};
-    std::map<int, std::shared_ptr<State>> trajectoryPredictionOtherVehicle{
+    std::map<size_t, std::shared_ptr<State>> trajectoryPredictionOtherVehicle{
             std::pair<int, std::shared_ptr<State>>(0, stateOtherZero),
             std::pair<int, std::shared_ptr<State>>(1, stateOtherOne)};
